refactor: format by styleguid
