+++ src/main/adapters/current-account-adapter.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '@/domain/entities'
+import { AccountModel } from '@/domain/models'
 import { makeLocalStorageAdapter } from '../factories/cache'
 
 export const setCurrentAccountAdapter = (account: AccountModel): void => {
